rm double inheritance and separate loss
